+++ CHANGELOG.md
@@ -43,6 +43,8 @@
   which those options have no effect ([rhbz#1533866])
 - When a user makes an error in a pcs command, usage for that specific command
   is printed instead of printing the whole usage
+- Show more user friendly error message when testing watchdog device and
+  multiple devices are present ([rhbz#1578891])
 
 ### Changed
 - Authentication has been overhauled ([rhbz#1549535]):